requirements.txt
streamlit==1.32.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
groq==0.4.1
pandas==2.2.1
numpy==1.26.4
pdfplumber==0.11.0
requests==2.31.0
python-dotenv==1.0.1
plotly==5.20.0
scikit-learn==1.4.1.post1
rank-bm25==0.2.2
